Don't use pip editable mode when installing deps (#67)
Not using -e anymore when installing p4-hlir from git. This was causing
issues when p4-hlir was also installed separately. Long story short, the
easy_install.pth file was pointing p4_hlir package to p4_hlir_v1_1 local
installation. Removing -e probably does not address the root casue of
the issue (which is the p4_hlir symlink in the p4v1.1 branch of the
p4-hlir repo), but it is an ok workaround for now.