Improve field list calculation support
More general support for fixed-size constants in field lists used as
input to calculations in P4. The bitwidth no longer needs to be a
multiple of 8. Added appropriate warnings and errors for violated
assumptions (except when a variable-length field is involved as we have
no visibility on the total bitwidth then).