Add pre-install hook to setup.py (#74)
This hook generates a wrapper script for p4c_bm.__main__.py which is
used in lieu of p4c-bmv2. The script adjust the PYTHONPATH dynamically
to include the installation path chosen py setuptools. For lazy users
who want to use a custom installation location but don't want to adjust
their PYHTONPATH themselves.