Reduce topo sorting run time (#80)
For more complex parser definitions with multiple tag stacks, the run
time of the algorithm which builds the header graph (in order to run the
topological sorting needed to program the deparser) was "exploding". To
fix this, I implemented a caching approach to avoid further processing
when the recursion visits the same parser state with the same processing
state (previous node + tag stacks indices) again. This reduced the run
time dramatically.